Progressively add files to public files cache
On large sites, the files table can be enourmous and it is
impractical to query the entire table for duplicates at once.

Instead, this patch modifies the public files check to query
for instances of a file individually (as required). Each time,
the content hash of that file is added to the cache and marked
as either public or private.